requirements.txt
einops
imageio
numpy<2
transformers
open_clip_torch==2.22.0
Pillow
tqdm
PyYAML
setuptools
scikit-learn
lightning
deepspeed
open3d
decord
omegaconf
opencv_python
pandas
moviepy
av<14
gradio
timm
kornia
fairscale
diffusers
httpx[socks]
tensorboard
accelerate
qwen_vl_utils[decord]
optimum
auto-gptq
autoawq
termcolor
wandb